поправлен stripe_checkout

+++ app/modules/orders/models.py
@@ -44,8 +44,8 @@ class Order(Base):
         nullable=False,
     )
 
-    total_amount: Mapped[int] = mapped_column(
-        Integer,
+    total_amount: Mapped[Decimal] = mapped_column(
+        Numeric(10, 2),
         nullable=False,
     )
